rename provider package to schema
