+++ main.go
@@ -2,12 +2,12 @@ package main
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/plugin"
-	"github.com/terraform-providers/terraform-provider-vault/provider"
+	"github.com/terraform-providers/terraform-provider-vault/schema"
 	"github.com/terraform-providers/terraform-provider-vault/vault"
 )
 
 func main() {
-	p := provider.New(vault.Provider())
+	p := schema.NewProvider(vault.Provider())
 	plugin.Serve(&plugin.ServeOpts{
 		ProviderFunc: p.ResourceProvider})
 }
